move some functions to infer
