+++ pyro/infer/util.py
@@ -28,30 +28,20 @@ def torch_backward(x):
         x.backward()
 
 
-def reduce_to_target(source, target):
-    """
-    Sums out any dimensions in source that are of size > 1 in source but of
-    size 1 in target.
-    """
-    while source.dim() > target.dim():
-        source = source.sum(0)
-    for k in range(1, 1 + source.dim()):
-        if source.size(-k) > target.size(-k):
-            source = source.sum(-k, keepdim=True)
-    return source
+def detach_iterable(iterable):
+    if torch.is_tensor(iterable):
+        return iterable.detach()
+    else:
+        return [var.detach() for var in iterable]
 
 
-def reduce_to_shape(source, shape):
+def zero_grads(tensors):
     """
-    Sums out any dimensions in source that are of size > 1 in source but of
-    size 1 in target.
+    Sets gradients of list of Variables to zero in place
     """
-    while source.dim() > len(shape):
-        source = source.sum(0)
-    for k in range(1, 1 + source.dim()):
-        if source.size(-k) > shape[-k]:
-            source = source.sum(-k, keepdim=True)
-    return source
+    for p in tensors:
+        if p.grad is not None:
+            p.grad = p.grad.new(p.shape).zero_()
 
 
 def get_iarange_stacks(trace):
